Feature/main character

+++ Assets/Interfaces/IAttacker.cs
@@ -1,6 +1,6 @@
 public interface IAttacker : IEntity
 {
-    int BaseAttack { get; }
+    float BaseAttack { get; }
     void Attack(AttackData attackData);
     AnimationController AnimationController { get; }
 }
